+++ src/main/java/org/magicdgs/readtools/tools/barcodes/BamBarcodeDetector.java
@@ -127,7 +127,7 @@ public final class BamBarcodeDetector extends ReadToolsBaseTool {
         progress.logNumberOfVariantsProcessed();
         decoder.logMatcherResult(logger);
         try {
-            decoder.outputStats(IOUtils.makeMetricsFile(outputPrefix));
+            decoder.outputStats(IOUtils.makeMetricsFile(outputPrefix).toFile());
         } catch (IOException e) {
             throw new UserException(e.getMessage(), e);
         }