+++ src/main/java/org/magicdgs/readtools/tools/barcodes/FastqBarcodeDetector.java
@@ -124,7 +124,7 @@ public final class FastqBarcodeDetector extends ReadToolsBaseTool {
         progress.logNumberOfVariantsProcessed();
         decoder.logMatcherResult(logger);
         try {
-            decoder.outputStats(IOUtils.makeMetricsFile(outputPrefix));
+            decoder.outputStats(IOUtils.makeMetricsFile(outputPrefix).toFile());
         } catch (IOException e) {
             throw new UserException(e.getMessage(), e);
         }